WIP commit, UL exec load libpcap.so
We perform function redirection on several libc functions in order
to have a dlopen() interface that can open our buffers as required.

This seems to be the most reliable approach.. it seems that the
vdso.so approach would not work - infact musl-libc couldn't load
libraries via the auxv vector, it would crash in sysv symbol lookup.